Fix Partner Pikachu box sprite
The pkuDexBuild-action merges arrays when resolving the $override command. As a result, the Partner Pikachu box sprites were not overriding the default form's but being appended onto them.

To fix this, the Front and Back sprite have been overridden explicitly.

A more permanent fix may involve adding options to the build workflow.